clean up and refactor module
