--- src/__init__.py
@@ -1,3 +1,0 @@
-import src.data
-import src.models
-import src.visualization
