+++ src/data/__init__.py
@@ -1,10 +1,7 @@
-from . import datasets
-from . import synthetic
-from . import utils
-from . import dset
-from . import localdata
-
-from .dset import Dataset
 from .datasets import *
+from .dset import *
+from .fetch import *
+from .localdata import *
+from .synthetic import *
+from .utils import *
 
-__all__ = (dset.__all__ + localdata.__all__)
