+++ src/data/localdata.py
@@ -9,8 +9,15 @@ from scipy.io import loadmat
 
 from .utils import read_space_delimited, normalize_labels
 
-__all__ = ['process_coil', 'process_frey_faces', 'process_hiva',
-           'process_lvq_pak', 'process_mnist', 'process_orl_faces', 'process_shuttle_statlog']
+__all__ = [
+    'process_coil',
+    'process_frey_faces',
+    'process_hiva',
+    'process_lvq_pak',
+    'process_mnist',
+    'process_orl_faces',
+    'process_shuttle_statlog',
+]
 
 logger = logging.getLogger(__name__)
